Set up search in other views
