+++ app/src/main/java/org/gophillygo/app/activities/EventsListActivity.java
@@ -1,16 +1,13 @@
 package org.gophillygo.app.activities;
 
-import android.app.SearchManager;
 import android.arch.lifecycle.LiveData;
 import android.arch.lifecycle.ViewModelProviders;
-import android.content.Context;
 import android.content.Intent;
 import android.databinding.DataBindingUtil;
 import android.os.Bundle;
 import android.support.annotation.NonNull;
 import android.support.v7.widget.LinearLayoutManager;
 import android.support.v7.widget.RecyclerView;
-import android.support.v7.widget.SearchView;
 import android.util.Log;
 import android.view.Menu;
 import android.view.MenuItem;
@@ -28,10 +25,6 @@ import org.gophillygo.app.data.networkresource.Status;
 import org.gophillygo.app.databinding.ActivityEventsListBinding;
 import org.gophillygo.app.databinding.FilterButtonBarBinding;
 import org.gophillygo.app.di.GpgViewModelFactory;
-import org.gophillygo.app.tasks.AddGeofencesBroadcastReceiver;
-import org.gophillygo.app.tasks.RemoveGeofenceWorker;
-
-import org.gophillygo.app.data.models.EventInfo;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -67,9 +60,7 @@ public class EventsListActivity extends FilterableListActivity
         // Get database ID for event clicked, based on positional offset, and pass it along
         long eventId = eventsListView.getAdapter().getItemId(position);
         Log.d(LOG_LABEL, "Clicked event with ID: " + eventId);
-        Intent intent = new Intent(this, EventDetailActivity.class);
-        intent.putExtra(EventDetailActivity.EVENT_ID_KEY, eventId);
-        startActivity(intent);
+        goToEvent(eventId);
     }
 
     public boolean clickedFlagOption(MenuItem item, AttractionInfo eventInfo, Integer position) {
@@ -142,12 +133,7 @@ public class EventsListActivity extends FilterableListActivity
     @Override
     public boolean onCreateOptionsMenu(Menu menu) {
         getMenuInflater().inflate(R.menu.events_list_menu, menu);
-
-        // Get the SearchView and set the searchable configuration
-        SearchManager searchManager = (SearchManager)getSystemService(Context.SEARCH_SERVICE);
-        SearchView searchView = (SearchView) menu.findItem(R.id.action_event_search).getActionView();
-        // Assumes current activity is the searchable activity
-        searchView.setSearchableInfo(searchManager.getSearchableInfo(getComponentName()));
+        setupSearch(menu, R.id.action_event_search);
         return true;
     }
