+++ app/src/main/java/org/gophillygo/app/activities/PlacesListActivity.java
@@ -1,14 +1,10 @@
 package org.gophillygo.app.activities;
 
-import android.app.SearchManager;
-import android.content.Context;
 import android.content.Intent;
-import android.database.Cursor;
 import android.databinding.DataBindingUtil;
 import android.os.Bundle;
 import android.support.annotation.NonNull;
 import android.support.annotation.Nullable;
-import android.support.v4.widget.CursorAdapter;
 import android.support.v7.widget.LinearLayoutManager;
 import android.support.v7.widget.RecyclerView;
 import android.support.v7.widget.SearchView;
@@ -36,8 +32,6 @@ import org.gophillygo.app.databinding.FilterButtonBarBinding;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
-
-import static android.app.SearchManager.SUGGEST_COLUMN_INTENT_DATA;
 
 public class PlacesListActivity extends FilterableListActivity implements
         PlacesListAdapter.AttractionListItemClickListener, SearchView.OnQueryTextListener {
@@ -71,36 +65,6 @@ public class PlacesListActivity extends FilterableListActivity implements
         // Get database ID for place clicked, based on positional offset, and pass it along
         long detailId = placesListView.getAdapter().getItemId(position);
         goToPlace(detailId);
-    }
-
-    private void goToAttractionForPosition(CursorAdapter adapter, int position) {
-        Log.d(LOG_LABEL, "onSuggestionClick " + position);
-        long itemId = adapter.getItemId(position);
-
-        Cursor cursor = (Cursor)adapter.getItem(position);
-        cursor.move(position);
-        int columnIndex = cursor.getColumnIndex(SUGGEST_COLUMN_INTENT_DATA);
-        int isEvent = cursor.getInt(columnIndex);
-
-        if (isEvent == 0) {
-            goToPlace(itemId);
-        } else {
-            goToEvent(itemId);
-        }
-    }
-
-    private void goToPlace(long detailId) {
-        Log.d(LOG_LABEL, "going to detail view for place ID " + detailId);
-        Intent intent = new Intent(this, PlaceDetailActivity.class);
-        intent.putExtra(PlaceDetailActivity.DESTINATION_ID_KEY, detailId);
-        startActivity(intent);
-    }
-
-    private void goToEvent(long detailId) {
-        Log.d(LOG_LABEL, "going to detail view for event ID " + detailId);
-        Intent intent = new Intent(this, EventDetailActivity.class);
-        intent.putExtra(EventDetailActivity.EVENT_ID_KEY, detailId);
-        startActivity(intent);
     }
 
     public boolean clickedFlagOption(MenuItem item, AttractionInfo destinationInfo, Integer position) {
@@ -191,31 +155,7 @@ public class PlacesListActivity extends FilterableListActivity implements
     @Override
     public boolean onCreateOptionsMenu(Menu menu) {
         getMenuInflater().inflate(R.menu.places_list_menu, menu);
-
-        // Get the SearchView and set the searchable configuration
-        SearchManager searchManager = (SearchManager)getSystemService(Context.SEARCH_SERVICE);
-        SearchView searchView = (SearchView) menu.findItem(R.id.action_place_list_search).getActionView();
-        // Assumes current activity is the searchable activity
-        searchView.setSearchableInfo(searchManager.getSearchableInfo(getComponentName()));
-
-        // handle opening detail intent from search
-        // https://developer.android.com/reference/android/widget/SearchView.OnSuggestionListener
-        searchView.setOnSuggestionListener(new SearchView.OnSuggestionListener() {
-            CursorAdapter adapter = searchView.getSuggestionsAdapter();
-
-            @Override
-            public boolean onSuggestionSelect(int position) {
-                Log.d(LOG_LABEL, "onSuggestionSelect " + position);
-                goToAttractionForPosition(adapter, position);
-                return true;
-            }
-
-            @Override
-            public boolean onSuggestionClick(int position) {
-                goToAttractionForPosition(adapter, position);
-                return true;
-            }
-        });
+        setupSearch(menu, R.id.action_place_list_search);
         return true;
     }
