+++ app/src/main/res/menu/home_menu.xml
@@ -2,7 +2,7 @@
 <menu xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto">
     <item
-        android:id="@+id/action_place_list_search"
+        android:id="@+id/action_home_search"
         android:orderInCategory="1"
         app:showAsAction="ifRoom"
         android:title="@string/menu_search_action"
